+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- markdown-include python package
+
 ## [2.0.1] - 2021-10-22
 
 ### Fixed